requirements.txt
# Base requirements (needed for both text and voice versions)
openai>=1.0.0
python-dotenv>=1.0.0

# Voice-specific requirements (only needed for solstis_voice.py)
elevenlabs>=0.2.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13  # Note: Requires portaudio to be installed first
Flask>=2.0.0  # For the web version 